+++ src/Timer.cu
@@ -19,7 +19,9 @@ Timer::Clock::duration Timer::get() {
   return end - start;
 }
 
-int Timer::getNs() {
+// TODO: this is prone to overflow: I might want to dynamically adjust
+// unit type / granularity as numbers get larger and larger
+int64_t Timer::getNs() {
   auto wall_time = std::chrono::duration_cast<std::chrono::nanoseconds>(
     end - start
   );
